chore: move kenmore logic to new DupScreenData.SpecialCases module
